fix(qr): mount scanner overlay once per layout
The "My QR" drawer kept rendering twice during scan because
WalletNavigation is rendered twice in each layout that owns it
(desktop sidebar + mobile bottom-nav slots in
src/app/(mobile-ui)/layout.tsx and
src/app/[...recipient]/payment-layout-wrapper.tsx). Each
WalletNavigation transitively rendered DirectSendQR's scanner +
drawer, which both portal to document.body via vaul/Radix, so the
portals stacked in body even when the parent was display:none.

The earlier snap-point tweaks (#2005, #2008, #2012) masked the
symptom but never fixed the structural duplication. Reverting
QRBottomDrawer to the original [0.75, 1] snap points and moving
the scanner/drawer/post-scan-modal logic into a new
QRScannerOverlay component that's mounted exactly once per layout
removes the duplicate mount entirely.

DirectSendQR is now a thin button that just flips
isQRScannerOpen on ModalsContext; the overlay listens to that
flag.
