+++ src/app/(mobile-ui)/layout.tsx
@@ -12,6 +12,7 @@ import { usePathname } from 'next/navigation'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { twMerge } from 'tailwind-merge'
 import '../../styles/globals.css'
+import QRScannerOverlay from '@/components/Global/QRScannerOverlay'
 import SupportDrawer from '@/components/Global/SupportDrawer'
 import JoinWaitlistPage from '@/components/Invites/JoinWaitlistPage'
 import { useRouter } from 'next/navigation'
@@ -222,6 +223,8 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
             <GuestLoginModal />
 
             <SupportDrawer />
+
+            <QRScannerOverlay />
         </div>
     )
 }
