+++ src/components/Global/DirectSendQR/index.tsx
@@ -1,480 +1,35 @@
 'use client'
-import { resolveEns } from '@/app/actions/ens'
+
 import { Button } from '@/components/0_Bruddle/Button'
-import Checkbox from '@/components/0_Bruddle/Checkbox'
-import { useToast } from '@/components/0_Bruddle/Toast'
-import Modal from '@/components/Global/Modal'
-import QRBottomDrawer from '@/components/Global/QRBottomDrawer'
-// QRScanner is NOT lazy-loaded - critical path for payments, needs instant response
-// 50KB bundle cost is worth it for better UX on primary flow
-import QRScanner from '@/components/Global/QRScanner'
-import { useAuth } from '@/context/authContext'
-import posthog from 'posthog-js'
-import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
-import * as Sentry from '@sentry/nextjs'
-import { usePathname, useRouter, useSearchParams } from 'next/navigation'
-import { useMemo, useState, type ChangeEvent } from 'react'
+import { useModalsContext } from '@/context/ModalsContext'
 import { twMerge } from 'tailwind-merge'
 import { type IconName } from '../Icons/Icon'
-import { EQrType, NAME_BY_QR_TYPE, parseEip681, recognizeQr } from './utils'
-import { pixKeyToBRCode } from '@/utils/pix.utils'
-import { useHaptic } from 'use-haptic'
-import { useModalsContext } from '@/context/ModalsContext'
-import { openExternalUrl } from '@/utils/capacitor'
-
-const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL!
-
-enum EModalType {
-    QR_NOT_SUPPORTED = 'QR_NOT_SUPPORTED',
-    WILL_BE_NOTIFIED = 'WILL_BE_NOTIFIED',
-    DIRECT_SEND = 'DIRECT_SEND',
-    EXTERNAL_URL = 'EXTERNAL_URL',
-    UNRECOGNIZED = 'UNRECOGNIZED',
-}
-
-type ModalType = `${EModalType}`
-
-interface ModalContentProps {
-    setModalContent: React.Dispatch<React.SetStateAction<ModalType | undefined>>
-    qrType: EQrType
-    setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>
-    redirectTo: string | undefined
-}
-
-const MODAL_CONTENTS: Record<ModalType, React.ComponentType<ModalContentProps>> = {
-    [EModalType.QR_NOT_SUPPORTED]: NotSupportedContent,
-    [EModalType.WILL_BE_NOTIFIED]: WillBeNotifiedContent,
-    [EModalType.DIRECT_SEND]: DirectSendContent,
-    [EModalType.EXTERNAL_URL]: ExternalUrlContent,
-    [EModalType.UNRECOGNIZED]: UnrecognizedContent,
-}
-
-// note: push notifications are now handled by onesignal via useNotifications hook.
-// this component just tracks the metric and shows confirmation.
-function NotSupportedContent({ setModalContent, qrType }: ModalContentProps) {
-    const { user } = useAuth()
-    return (
-        <div className="flex flex-col justify-center p-6">
-            <span className="text-sm">We're working on an integration.</span>
-            <span className="text-sm">Get notified when it goes live!</span>
-            <Button
-                onClick={() => {
-                    setModalContent(EModalType.WILL_BE_NOTIFIED)
-                    posthog.capture(ANALYTICS_EVENTS.QR_NOTIFY_ME_CLICKED, { qr_type: qrType })
-                }}
-                className="mt-4 w-full"
-                shadowType="primary"
-                shadowSize="4"
-            >
-                Get notified!
-            </Button>
-        </div>
-    )
-}
-
-function WillBeNotifiedContent({ qrType, setIsModalOpen }: ModalContentProps) {
-    return (
-        <div className="flex flex-col justify-center p-6">
-            <span className="text-sm">We'll let you know as soon as {NAME_BY_QR_TYPE[qrType]} is supported.</span>
-            <Button
-                onClick={() => setIsModalOpen(false)}
-                className="mt-4 w-full"
-                variant="primary-soft"
-                shadowType="primary"
-                shadowSize="4"
-            >
-                Close
-            </Button>
-        </div>
-    )
-}
-
-function DirectSendContent({ redirectTo, setIsModalOpen }: ModalContentProps) {
-    const [userAcknowledged, setUserAcknowledged] = useState(false)
-    const router = useRouter()
-    return (
-        <div className="flex flex-col justify-center p-6">
-            <span className="text-sm">Peanut supports cross-chain payments.</span>
-            <span className="text-sm">Please confirm the payment details before sending</span>
-            <Checkbox
-                value={userAcknowledged}
-                onChange={(e: ChangeEvent<HTMLInputElement>) => {
-                    setUserAcknowledged(e.target.checked)
-                }}
-                className="mt-4"
-                label="I understand and will confirm payment details."
-            />
-            <Button
-                onClick={() => {
-                    router.push(redirectTo!)
-                    setTimeout(() => {
-                        setIsModalOpen(false)
-                    }, 750)
-                }}
-                disabled={!userAcknowledged}
-                className="mt-4 w-full"
-                shadowType="primary"
-                shadowSize="4"
-            >
-                Continue
-            </Button>
-        </div>
-    )
-}
-
-function ExternalUrlContent({ redirectTo, setIsModalOpen }: ModalContentProps) {
-    return (
-        <div className="flex flex-col justify-center p-6">
-            <span className="text-sm">Peanut doesn't support this QR but you can open it with your browser. </span>
-            <span className="text-sm">Make sure you trust this website!</span>
-            <div className="flex items-center justify-center gap-2">
-                <Button
-                    onClick={() => {
-                        if (redirectTo) openExternalUrl(redirectTo)
-                        setTimeout(() => {
-                            setIsModalOpen(false)
-                        }, 750)
-                    }}
-                    className="mt-4 w-full"
-                    shadowType="primary"
-                    shadowSize="4"
-                >
-                    Open link
-                </Button>
-                <Button
-                    onClick={() => setIsModalOpen(false)}
-                    className="mt-4 w-full"
-                    variant="primary-soft"
-                    shadowType="primary"
-                    shadowSize="4"
-                >
-                    Close
-                </Button>
-            </div>
-        </div>
-    )
-}
-
-function UnrecognizedContent({ setIsModalOpen }: ModalContentProps) {
-    return (
-        <div className="flex flex-col justify-center p-6">
-            <span className="text-sm">Sorry, this QR code couldn't be recognized.</span>
-            <Button onClick={() => setIsModalOpen(false)} className="mt-4 w-full" shadowType="primary" shadowSize="4">
-                Okay
-            </Button>
-        </div>
-    )
-}
 
 export default function DirectSendQr({
     icon = 'qr-code',
     className = '',
-    ctaTitle,
     disabled = false,
 }: {
     className?: string
-    ctaTitle?: string
     icon?: IconName
     disabled?: boolean
 }) {
-    const [isModalOpen, setIsModalOpen] = useState(false)
-    const [qrType, setQrType] = useState<EQrType | undefined>(undefined)
-    const [redirectTo, setRedirectTo] = useState<string | undefined>(undefined)
-    const [modalContent, setModalContent] = useState<ModalType | undefined>(undefined)
-    const router = useRouter()
-    const pathname = usePathname()
-    const searchParams = useSearchParams()
-    const toast = useToast()
-    const { user } = useAuth()
-    const payUserUrl = useMemo(() => {
-        if (!user?.user.username) return ''
-        return `${BASE_URL}/pay/${user.user.username}`
-    }, [user?.user.username])
-    const { triggerHaptic } = useHaptic()
-    const { isQRScannerOpen, setIsQRScannerOpen } = useModalsContext()
-
-    const startScanner = () => {
-        setIsQRScannerOpen(true)
-    }
-
-    const processQRCode = async (data: string): Promise<{ success: boolean; error?: string }> => {
-        triggerHaptic()
-
-        let redirectUrl: string | undefined = undefined
-        let toConfirmUrl: string | undefined = undefined
-        const originalData = data
-        data = data.toLowerCase()
-        const qrType = recognizeQr(data)
-
-        // Redact sensitive data for privacy before logging to analytics
-        const getLogData = () => {
-            // PIX keys contain PII (CPF, phone, email) - only log the key type
-            if (qrType === EQrType.PIX_KEY) {
-                const trimmed = originalData.trim()
-                if (trimmed.startsWith('+') || /^55\d/.test(trimmed)) return 'pix:phone'
-                if (/^\d{11}$/.test(trimmed)) return 'pix:cpf'
-                if (/^\d{14}$/.test(trimmed)) return 'pix:cnpj'
-                if (trimmed.includes('@')) return 'pix:email'
-                if (/^[0-9a-f]{8}-[0-9a-f]{4}-/i.test(trimmed)) return 'pix:uuid'
-                return 'pix:unknown'
-            }
-            // Claim links contain private keys - redact the sensitive part
-            if (qrType === EQrType.PEANUT_URL && originalData.toLowerCase().includes('/claim')) {
-                return 'peanut:claim-link'
-            }
-            return originalData
-        }
-        posthog.capture(ANALYTICS_EVENTS.QR_SCANNED, { qr_type: qrType, data: getLogData() })
-        setQrType(qrType as EQrType)
-        switch (qrType) {
-            case EQrType.PEANUT_URL:
-                {
-                    // Extract path by removing domain (handle with/without protocol and www)
-                    let path = originalData
-                        .replace(/^https?:\/\/(www\.)?/, '') // Remove protocol and www
-                        .replace(/^[^/]+/, '') // Remove domain
-
-                    if (!path.startsWith('/')) {
-                        path = '/' + path
-                    }
-
-                    // Special handling for /qr/ paths (redirect QR codes - user-tied QRs)
-                    if (path.startsWith('/qr/')) {
-                        const redirectQrCode = path.substring(4) // Remove '/qr/' prefix
-
-                        // Check redirect QR status (single endpoint for speed)
-                        try {
-                            const { serverFetch } = await import('@/utils/api-fetch')
-                            const response = await serverFetch(`/qr/${redirectQrCode}`, {
-                                method: 'GET',
-                            })
-                            const data = await response.json()
-
-                            if (data.claimed && data.redirectUrl) {
-                                // Redirect QR is claimed, redirect to target URL
-                                redirectUrl = data.redirectUrl
-                            } else {
-                                // Redirect QR is available, go to claim/redirect page
-                                redirectUrl = `/qr/${redirectQrCode}`
-                            }
-                        } catch (error) {
-                            console.error('Error checking redirect QR:', error)
-                            // Fallback: redirect to claim page
-                            redirectUrl = `/qr/${redirectQrCode}`
-                        }
-                    } else {
-                        redirectUrl = path
-                    }
-                }
-                break
-            case EQrType.EVM_ADDRESS:
-                {
-                    toConfirmUrl = `/${data}`
-                }
-                break
-            case EQrType.EIP_681:
-                {
-                    try {
-                        const { address, chainId, amount, tokenSymbol } = parseEip681(data)
-                        toConfirmUrl = `/${address}`
-                        if (chainId) {
-                            toConfirmUrl += `@${chainId}`
-                            if (tokenSymbol) {
-                                toConfirmUrl += `/`
-                                if (amount) {
-                                    toConfirmUrl += `${amount}`
-                                }
-                                toConfirmUrl += `${tokenSymbol}`
-                            }
-                        }
-                    } catch (error) {
-                        toast.error('Error parsing EIP-681 URL')
-                        Sentry.captureException(error)
-                    }
-                }
-                break
-            case EQrType.ENS_NAME:
-                {
-                    const resolvedAddress = await resolveEns(data)
-                    if (!!resolvedAddress) {
-                        toConfirmUrl = `/${data}`
-                    }
-                }
-                break
-            case EQrType.MERCADO_PAGO:
-            case EQrType.ARGENTINA_QR3:
-            case EQrType.PIX:
-                {
-                    const timestamp = Date.now()
-                    // Casing matters, so send original instead of normalized
-                    redirectUrl = `/qr-pay?qrCode=${encodeURIComponent(originalData)}&t=${timestamp}&type=${qrType}`
-                }
-                break
-            case EQrType.PIX_KEY:
-                {
-                    const brCode = pixKeyToBRCode(originalData)
-                    if (brCode) {
-                        const timestamp = Date.now()
-                        redirectUrl = `/qr-pay?qrCode=${encodeURIComponent(brCode)}&t=${timestamp}&type=${EQrType.PIX}`
-                    } else {
-                        // Invalid PIX key - show unrecognized modal
-                        setModalContent(EModalType.UNRECOGNIZED)
-                        setIsModalOpen(true)
-                        setIsQRScannerOpen(false)
-                        return { success: true }
-                    }
-                }
-                break
-            case EQrType.BITCOIN_ONCHAIN:
-            case EQrType.BITCOIN_INVOICE:
-            case EQrType.TRON_ADDRESS:
-            case EQrType.SOLANA_ADDRESS:
-            case EQrType.XRP_ADDRESS: {
-                setModalContent(EModalType.QR_NOT_SUPPORTED)
-                setIsModalOpen(true)
-                setIsQRScannerOpen(false)
-                return { success: true }
-            }
-            case EQrType.URL: {
-                setRedirectTo(originalData)
-                setModalContent(EModalType.EXTERNAL_URL)
-                setIsModalOpen(true)
-                setIsQRScannerOpen(false)
-                return { success: true }
-            }
-            default:
-                {
-                    setModalContent(EModalType.UNRECOGNIZED)
-                    setIsModalOpen(true)
-                    setIsQRScannerOpen(false)
-                }
-                break
-        }
-
-        if (redirectUrl) {
-            const currentSearchParams = searchParams.toString()
-            let currentFullPath = pathname
-            currentFullPath = currentSearchParams ? `${currentFullPath}?${currentSearchParams}` : currentFullPath
-            currentFullPath += window.location.hash
-
-            if (
-                currentFullPath === redirectUrl ||
-                (redirectUrl.startsWith('/') && currentFullPath === redirectUrl.substring(1))
-            ) {
-                // We're already at this location, just close the scanner
-                setIsQRScannerOpen(false)
-            } else {
-                router.push(redirectUrl)
-                setIsQRScannerOpen(false)
-            }
-            return { success: true }
-        }
-
-        if (toConfirmUrl) {
-            setModalContent(EModalType.DIRECT_SEND)
-            setIsModalOpen(true)
-            setIsQRScannerOpen(false)
-            setRedirectTo(toConfirmUrl)
-            return { success: true }
-        }
-
-        // close the scanner and show the error toast
-        setIsQRScannerOpen(false)
-
-        return {
-            success: false,
-            error: 'QR not recognized as Peanut URL',
-        }
-    }
-
-    const modalTitle = useMemo(() => {
-        let title: string | undefined
-        if (modalContent && qrType) {
-            switch (modalContent) {
-                case EModalType.QR_NOT_SUPPORTED:
-                    {
-                        const qrTypeName = NAME_BY_QR_TYPE[qrType]
-                        title = `${qrTypeName} not supported yet.`
-                    }
-                    break
-                case EModalType.WILL_BE_NOTIFIED:
-                    {
-                        title = "You're on the list!"
-                    }
-                    break
-                case EModalType.DIRECT_SEND:
-                    {
-                        title = 'ℹ️ Payment Confirmation'
-                    }
-                    break
-                case EModalType.EXTERNAL_URL:
-                    {
-                        title = 'This is an external link!'
-                    }
-                    break
-            }
-        } else if (modalContent === EModalType.UNRECOGNIZED) {
-            title = 'Unrecognized QR code'
-        }
-        return title
-    }, [modalContent, qrType])
+    const { setIsQRScannerOpen } = useModalsContext()
 
     return (
-        <>
-            <Button
-                onClick={startScanner}
-                variant="purple"
-                shadowSize="4"
-                shadowType="primary"
-                className={twMerge(
-                    'mx-auto h-20 w-20 cursor-pointer justify-center rounded-full p-4.5 hover:bg-primary-1/100 active:translate-x-0 active:translate-y-0 active:scale-95',
-                    className
-                )}
-                disabled={disabled}
-                icon={icon}
-                iconSize={32}
-                iconContainerClassName="!size-8"
-            />
-
-            <Modal
-                title={modalTitle}
-                visible={isModalOpen && !!modalContent}
-                onClose={() => setIsModalOpen(false)}
-                initialFocus={undefined}
-                preventClose={modalContent !== EModalType.QR_NOT_SUPPORTED}
-                className="items-center rounded-none"
-                classWrap="sm:m-auto sm:self-center self-center m-4 bg-background rounded-none border-0"
-                classButtonClose="hidden"
-            >
-                {modalContent &&
-                    (() => {
-                        const ModalComponent = MODAL_CONTENTS[modalContent]
-                        return (
-                            <ModalComponent
-                                setModalContent={setModalContent}
-                                qrType={qrType!}
-                                setIsModalOpen={setIsModalOpen}
-                                redirectTo={redirectTo}
-                            />
-                        )
-                    })()}
-            </Modal>
-
-            {isQRScannerOpen && (
-                <>
-                    <QRScanner onScan={processQRCode} onClose={() => setIsQRScannerOpen(false)} isOpen={true} />
-                    {/* Render QRBottomDrawer with z-[60] to ensure it appears above QRScanner (z-50)
-                        This allows "scan OR be scanned" dual functionality */}
-                    <QRBottomDrawer
-                        url={payUserUrl}
-                        collapsedTitle="My QR"
-                        expandedTitle="Show QR to Get Paid"
-                        text="Let others scan this to pay you"
-                        buttonText="Share your profile"
-                        className="z-[60]"
-                    />
-                </>
+        <Button
+            onClick={() => setIsQRScannerOpen(true)}
+            variant="purple"
+            shadowSize="4"
+            shadowType="primary"
+            className={twMerge(
+                'mx-auto h-20 w-20 cursor-pointer justify-center rounded-full p-4.5 hover:bg-primary-1/100 active:translate-x-0 active:translate-y-0 active:scale-95',
+                className
             )}
-        </>
+            disabled={disabled}
+            icon={icon}
+            iconSize={32}
+            iconContainerClassName="!size-8"
+        />
     )
 }
